fix: Depend on sdl2 and sdl2_ttf provided by system as default
As default, meson will use sdl2 and sdl2_ttf provided by system
if available. If they are not provided, or if the
use_sdl2_subproject build option is true, sdl2 and sdl2_ttf will
be built as subprojects. The main motivation for building them
as subprojects is if the system provides an outdated version
of sdl2, ie < 2.24.0.

Signed-off-by: Gustav Grusell <gustav.grusell@gmail.com>