fix(imgui): stepping backwards now works correctly
In some circumstances, for instance if steeping back one frame from an iframe
Seeking may not work as expected due to av_seek_frame apperantly seeking on packet dts
Which may be slightly lower than seek dts we calculate. Therefore, we retry seek
with seeking to an earlier position in stream.

Signed-off-by: Gustav Grusell <gustav.grusell@gmail.com>